backend/ai/requirements.txt
fastapi
uvicorn
opencv-python
numpy
pillow
mediapipe
moviepy
rembg
python-multipart
